Arreglo de HATEOAS y testing de controller

+++ productservice/productservice/src/main/java/com/perfulandia/productservice/service/ProductoService.java
@@ -1,7 +1,9 @@
 package com.perfulandia.productservice.service;
 
+import com.perfulandia.productservice.exception.ResourceNotFoundException;
 import com.perfulandia.productservice.model.Producto;
 import com.perfulandia.productservice.repository.ProductoRepository;
+
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -23,12 +25,16 @@ public class ProductoService {
     public Producto guardar(Producto producto){
         return repo.save(producto);
     }
-    //Buscar
-    public Producto bucarPorId(long id){
-        return repo.findById(id).orElse(null);
+    //Buscar por id
+    public Producto buscarPorId(long id){
+        return repo.findById(id)
+                .orElseThrow(() -> new ResourceNotFoundException("Producto no encontrado con ID: " + id));
     }
-    //Eliminar
+    //Eliminar por id
     public void eliminar(long id){
+        if (!repo.existsById(id)) {
+            throw new ResourceNotFoundException("No se puede eliminar. Usuario no encontrado con ID: " + id);
+        }
         repo.deleteById(id);
     }
 }
